requirements.txt
streamlit
pandas
numpy
yfinance
plotly
pykrx
google-generativeai